-Fix: 32/64bit related compilation warning.
git-svn-id: http://svn.fuzzle.org/mloop/trunk@25 ba049829-c6ef-42ef-81ac-908dd8d2e907

+++ src/loop.cpp
@@ -152,12 +152,12 @@ void Loop::Empty()
 void Loop::Save(FILE *f) const
 {
 	fprintf(f, "%u %f %f\n", m_length, m_position, m_tempo);
-	fprintf(f, "%lu\n", m_events.size());
+	fprintf(f, "%lu\n", (unsigned long)m_events.size());
 
 	EventList::const_iterator it;
 	for (it = m_events.begin(); it != m_events.end(); ++it) {
 		const jack_midi_event_t &ev = *it;
-		fprintf(f, "%u %lu", ev.time, ev.size);
+		fprintf(f, "%u %lu", ev.time, (unsigned long)ev.size);
 		for (uint i = 0; i < ev.size; i++) {
 			fprintf(f, " %02X", ev.buffer[i]);
 		}
